Implemented A* search
It is very similar to uniform cost search. The only change is the priority we assign when enqueuing a state. instead of just using the cumulative cost from the start to the current state (g-value), we also add the value determined by a heuristic function that calculates a cost from the current state to the goal state (h-value). the sum of the h-value and the g-value will make up the f-value which will be the priority we will queue a certain state with.